Move command registration
