+++ public/script.js
@@ -212,7 +212,7 @@ import {
     selectContextPreset,
 } from './scripts/instruct-mode.js';
 import { initLocales, t, translate } from './scripts/i18n.js';
-import { getAvailableTokenizers, getFriendlyTokenizerName, getTokenCount, getTokenCountAsync, getTokenizerModel, initTokenizers, saveTokenCache, selectTokenizer } from './scripts/tokenizers.js';
+import { getFriendlyTokenizerName, getTokenCount, getTokenCountAsync, getTokenizerModel, initTokenizers, saveTokenCache } from './scripts/tokenizers.js';
 import {
     user_avatar,
     getUserAvatars,
@@ -8451,27 +8451,6 @@ async function selectInstructCallback(_, name) {
     return foundName;
 }
 
-async function selectTokenizerCallback(_, name) {
-    if (!name) {
-        return getAvailableTokenizers().find(tokenizer => tokenizer.tokenizerId === power_user.tokenizer)?.tokenizerKey ?? '';
-    }
-
-    const tokenizers = getAvailableTokenizers();
-    const fuse = new Fuse(tokenizers, { keys: ['tokenizerKey', 'tokenizerName'] });
-    const result = fuse.search(name);
-
-    if (result.length === 0) {
-        toastr.warning(`Tokenizer "${name}" not found`);
-        return '';
-    }
-
-    /** @type {import('./scripts/tokenizers.js').Tokenizer} */
-    const foundTokenizer = result[0].item;
-    selectTokenizer(foundTokenizer.tokenizerId);
-
-    return foundTokenizer.tokenizerKey;
-}
-
 async function enableInstructCallback() {
     $('#instruct_enabled').prop('checked', true).trigger('change');
     return '';
@@ -9115,28 +9094,6 @@ jQuery(async function () {
                 </ul>
             </div>
         `,
-    }));
-    SlashCommandParser.addCommandObject(SlashCommand.fromProps({
-        name: 'tokenizer',
-        callback: selectTokenizerCallback,
-        returns: 'current tokenizer',
-        unnamedArgumentList: [
-            SlashCommandArgument.fromProps({
-                description: 'tokenizer name',
-                typeList: [ARGUMENT_TYPE.STRING],
-                enumList: getAvailableTokenizers().map(tokenizer =>
-                    new SlashCommandEnumValue(tokenizer.tokenizerKey, tokenizer.tokenizerName, enumTypes.enum, enumIcons.default)),
-            }),
-        ],
-        helpString: `
-            <div>
-                Selects tokenizer by name. Gets the current tokenizer if no name is provided.
-            </div>
-            <div>
-                <strong>Available tokenizers:</strong>
-                <pre><code>${getAvailableTokenizers().map(t => t.tokenizerKey).join(', ')}</code></pre>
-            </div>
-        `
     }));
     SlashCommandParser.addCommandObject(SlashCommand.fromProps({
         name: 'instruct-on',
